requirements.txt
av==12.3.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==4.4.0
faster-whisper==1.0.3
filelock==3.13.1
flatbuffers==24.3.25
fsspec==2024.2.0
huggingface-hub==0.26.0
humanfriendly==10.0
idna==3.10
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==2.1.2
onnxruntime==1.19.2
packaging==24.1
protobuf==5.28.2
PyAudio==0.2.14
pyreadline3==3.5.4
PySide6==6.8.0.1
PySide6_Addons==6.8.0.1
PySide6_Essentials==6.8.0.1
PyYAML==6.0.2
requests==2.32.3
shiboken6==6.8.0.1
sympy==1.12
tokenizers==0.20.1
tqdm==4.66.5
typing_extensions==4.9.0
urllib3==2.2.3
